Isolate ArrayAccess from Model to optional trait
